Optimized Select And Poll Socket Interfaces
The host can now obtain a handle to the underlying socket objects. As such, the socket-specific poll and select functions can now accept multiple handles in one location independent of the object. A future speed boost may be to add caching so the SocketPollTokens don't need to be copied between pollfd every call.